Decode u64 relation masks, pattern payloads, and EntityGenesis headers
Text-frame relations carry constraint bit 40, which the u32 mask read
truncated, and streams in the `EntityGenesis` header form decoded no
sketch entity headers at all, leaving every sketch point, curve, and
relation unbound. The mask width follows the record dialect: relations
with an `EntityGenesis` block store an unmarked u64 six zero bytes after
the owner reference, while records without it keep the marked or direct
u32 form, so both dialects decode without ambiguity.

Widen `SketchRelation::state` and `unknown_constraint_bits` to u64, add
the `TextFrame`/`TextPath` kinds, and rework `parse_sketch_relation` to
read per-member u32 roles, the `EntityGenesis` origin bitfield, and the
dialect-correct mask position. Add `decode_pattern_definition` for the
circular, rectangular, and text-frame auxiliary layouts, teach
`decode_entity_headers` the `EntityGenesis` header variant via
`parse_genesis_entity_header`, and make the writer emit the genesis block
plus padded u64 mask for states above u32 range while patching mask edits
at their source width. Cover the new grammar with synthetic-fixture tests
for the header variant, the u64 text-frame mask, member roles, and both
pattern auxiliary runs.

+++ crates/cadmpeg-codec-f3d/src/lib.rs
@@ -1852,7 +1852,8 @@ pub fn validate_native(ir: &CadIr) -> Vec<Finding> {
         let mut return_members = group.return_members.clone();
         locus_members.sort_unstable();
         return_members.sort_unstable();
-        let (expected_kinds, expected_unknown) = design::decode_constraint_kinds(group.state);
+        let (expected_kinds, expected_unknown) =
+            design::decode_constraint_kinds(u64::from(group.state));
         let owner_is_sketch = entities_by_suffix
             .get(&(native_stream, u64::from(group.owner_reference)))
             .is_some_and(|entity| entity.object_kind == Some(records::DesignObjectKind::Sketch));
@@ -1881,7 +1882,7 @@ pub fn validate_native(ir: &CadIr) -> Vec<Finding> {
             && returns_valid
             && locus_members == return_members
             && group.constraint_kinds == expected_kinds
-            && group.unknown_constraint_bits == expected_unknown
+            && u64::from(group.unknown_constraint_bits) == expected_unknown
             && group.next_byte_offset
                 == returns_start
                     .saturating_add((count as u64).saturating_mul(11))
